move Input to common folder
